2021-10-24 Devide User Modal Section && Add LogOut Function
