+++ src/Components/Location/UserProfile.js
@@ -1,8 +1,7 @@
-import React, { useState, useContext } from 'react';
-import { Link } from 'react-router-dom';
-import { ListContext } from 'Routers/Router';
+import React, { useState } from 'react';
 import styled from 'styled-components';
 import userIcon from 'assets/userIcon.jpeg';
+import UserModal from './UserModal';
 
 const LayerProfile = styled.div`
     width: 40px;
@@ -17,69 +16,8 @@ const LayerProfile = styled.div`
     border-radius: 50%;
 `;
 
-const LayerUserInfo = styled.div`
-    width: 300px;
-    height: 100px;
-    background-color: white;
-    position: absolute;
-    top: 8%;
-    right: 1%;
-    z-index: 1;
-    border: 1px solid rgba(10, 10, 10, .35);
-    border-radius: 10px;
-    display: flex;
-`;
-
-const LayerUser = styled.div`
-    width: 33%;
-    height:80%;
-`;
-
-const Image = styled.img`
-    width: 70%;
-    height: 80%;
-    margin: 10px 0 0 20px;
-`;
-
-const UserName = styled.h3`
-    font-size: 1.7rem;
-    margin: 20px 0 10px 0;
-`;
-
-const UserEditLink = styled(Link)`
-    font-size: 1.2rem;
-`;
-
-const LogOutButton = styled.button`
-    all:unset;
-    width: 60px;
-    height: 25px;
-    position: absolute;
-    top: 15%;
-    right: 8%;
-    font-size: 1.4rem;
-    text-align: center;
-    border:1px solid rgba(10, 10, 10, .35);
-`;
-
-const UserInfo = ({userInfo}) => (
-    <>
-        <LayerUserInfo>
-            <LayerUser>
-                <Image src={userIcon} alt="Profile_Image"/>
-            </LayerUser>
-            <LayerUser>
-                <UserName>{userInfo.name}</UserName>
-                <UserEditLink to="/users"> 프로필 수정 </UserEditLink>
-            </LayerUser>
-            <LogOutButton type="submit"> 로그아웃 </LogOutButton>
-        </LayerUserInfo>
-    </>
-)
-
 const UserProfile = () => {
     const [userProfile, setUserProfile] = useState(false);
-    const {userInfo} = useContext(ListContext);
 
     const handleUserModal = evt => {
         setUserProfile(!userProfile);
@@ -88,7 +26,7 @@ const UserProfile = () => {
     return (
         <>
             <LayerProfile onClick={handleUserModal} />
-            {userProfile && <UserInfo userInfo={userInfo}/>}
+            {userProfile && <UserModal />}
         </>
     )
 }
